requirements.txt
paramiko==3.4.0
markdown-it-py==3.0.0
pyTelegramBotAPI==4.15.2
python-telegram-bot==20.7
requests==2.31.0